positive/negative effect potion mod count can be a float. In the rare case, selected additional effects could be ones already present on the potion. Will now be rolled up to 5 times

+++ src/main/java/de/dafuqs/spectrum/recipe/potion_workshop/PotionMod.java
@@ -10,8 +10,8 @@ public class PotionMod {
 	public float flatPotencyBonusPositiveEffects = 0.0F;
 	public float flatPotencyBonusNegativeEffects = 0.0F;
 	
-	public int additionalRandomPositiveEffectCount = 0;
-	public int additionalRandomNegativeEffectCount = 0;
+	public float additionalRandomPositiveEffectCount = 0;
+	public float additionalRandomNegativeEffectCount = 0;
 	
 	public float chanceToAddLastEffect = 0.0F;
 	public float lastEffectPotencyModifier = 1.0F;
